+++ include/CQChartsSankeyPlot.h
@@ -329,6 +329,9 @@ class CQChartsSankeyPlotEdge {
   const OptReal &value() const { return value_; }
   void setValue(const OptReal &r) { value_ = r; }
 
+  const CQChartsColumn &valueColumn() const { return valueColumn_; }
+  void setValueColumn(const CQChartsColumn &c) { valueColumn_ = c; }
+
 #if 0
   //! get/set label
   const QString &label() const { return label_; }
@@ -375,17 +378,18 @@ class CQChartsSankeyPlotEdge {
   bool edgePath(QPainterPath &path, bool isLine=false) const;
 
  protected:
-  const Plot*   plot_     { nullptr }; //!< plot
-  int           id_       { -1 };      //!< unique id
-  OptReal       value_;                //!< value
-//QString       label_;                //!< label
-  CQChartsColor color_;                //!< color
-  int           pathId_   { -1 };      //!< path id
-  ModelInds     modelInds_;            //!< model inds
-  Node*         srcNode_  { nullptr }; //!< source node
-  Node*         destNode_ { nullptr }; //!< destination node
-  Obj*          obj_      { nullptr }; //!< associated edge object
-  bool          isLine_   { false };   //!< is edge a line
+  const Plot*    plot_     { nullptr }; //!< plot
+  int            id_       { -1 };      //!< unique id
+  OptReal        value_;                //!< value
+  CQChartsColumn valueColumn_;          //!< value column
+//QString        label_;                //!< label
+  CQChartsColor  color_;                //!< color
+  int            pathId_   { -1 };      //!< path id
+  ModelInds      modelInds_;            //!< model inds
+  Node*          srcNode_  { nullptr }; //!< source node
+  Node*          destNode_ { nullptr }; //!< destination node
+  Obj*           obj_      { nullptr }; //!< associated edge object
+  bool           isLine_   { false };   //!< is edge a line
 };
 
 //---